+++ lib/src/features/home_store/domain/sales_item_data.dart
@@ -1,0 +1,11 @@
+class SalesItem {
+  final String imageUrl;
+  final String pageTitle;
+  final String pageDescription;
+
+  const SalesItem({
+    required this.imageUrl,
+    required this.pageTitle,
+    required this.pageDescription,
+  });
+}
